% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{apply_defaults}
\alias{apply_defaults}
\title{Ensure Valid Inputs}
\usage{
apply_defaults(parameters)
}
\arguments{
\item{parameters}{A list}
}
\value{
An updated list
}
\description{
Fills in any missing inputs with typical values.
}
\details{
Takes an existing list and adds tags and default values for missing items.
}
