% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{extract_individuals}
\alias{extract_individuals}
\title{Retrieve Results}
\usage{
extract_individuals(pop_gen_out, convert = TRUE)
}
\arguments{
\item{pop_gen_out}{Structure returned from \code{\link{generate_population}}}

\item{convert}{Whether to convert service output to numerical data, and convert to a data frame}
}
\value{
If convert is TRUE, a data frame; otherwise, a matrix
}
\description{
Parse the service output for population data.
}
\details{
When converting numerical data, any NAs produced are not accompanied by a diagnostic message.
}
