% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{generate_failed}
\alias{generate_failed}
\title{Query Output}
\usage{
generate_failed(pop_gen_out)
}
\arguments{
\item{pop_gen_out}{Structure returned from \code{\link{generate_population}}}
}
\value{
TRUE if pop_gen_out does not contain population data, FALSE otherwise
}
\description{
Returns TRUE if the call failed to generate population data.
}
