% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpgcli.R
\name{generate_population}
\alias{generate_population}
\title{Call PopGen Web Service}
\usage{
generate_population(parameters, silent = FALSE)
}
\arguments{
\item{parameters}{Inputs that describe the target population}

\item{silent}{Whether or not to print activity messages during the call}
}
\value{
A structure containing data returned by the PopGen service.
}
\description{
Sample a population dataset using PopGen.
}
\details{
Use of this API requires an active account on the \href{http://xnet.hsl.gov.uk/popgen}{PopGen} web site.

The returned object consists either of an XML document fragment containing the generated population, or one or more error messages.
}
\examples{
# Set PopGen account name
parameters <- list(
  PopGenUserName = "kevin.soandso@some.sci.org",
  PopulationSize = 100
  )

# Add typical values for other inputs
parameters <- apply_defaults(parameters)

# Call web service
pop_gen_out <- generate_population(parameters)
}
\seealso{
Query outcome of service call with \code{\link{generate_succeeded}}. Retrieve results with \code{\link{extract_individuals}}.
}
