% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{generate_succeeded}
\alias{generate_succeeded}
\title{Query Output}
\usage{
generate_succeeded(pop_gen_out)
}
\arguments{
\item{pop_gen_out}{Structure returned from \code{\link{generate_population}}}
}
\value{
TRUE if pop_gen_out contains population data, FALSE otherwise
}
\description{
Returns TRUE if the call succeeded in generating population data.
}
