% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_default_parameters}
\alias{get_default_parameters}
\title{List of Inputs}
\usage{
get_default_parameters()
}
\value{
A list of required inputs
}
\description{
Get a list of required inputs with typical settings.
}
\details{
PopGenUserName must be assigned the email address of an active \href{http://xnet.hsl.gov.uk/popgen}{PopGen} account.
}
\examples{
parameters <- get_default_parameters()
parameters$InVitroEnzymeRates = list(
  c(Enzyme = "CYP1A2", Rate = 123),
  c(Enzyme = "CYP2B6", Rate = 456)
  )
}
