% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpgcli.R
\docType{package}
\name{Rpgcli}
\alias{Rpgcli}
\alias{Rpgcli-package}
\title{PopGen Client}
\description{
A set of R functions for calling the \href{http://xnet.hsl.gov.uk/popgen}{PopGen} web service and for extracting data from the response.
}
\details{
Pay a visit to \href{http://xnet.hsl.gov.uk/popgen}{PopGen} and review the inputs form before
using this package for programmatic access.

Refer to the service's \href{http://xnet.hsl.gov.uk/Popgen/svc/pg.svc?xsd=xsd2}{inputs schema} for concise
coverage of the input requirements.
}
\examples{
# this example samples data from all four datasets
datasets <- c("P3M", "ICRP", "HSE", "NDNS")
n_datasets <- length(datasets)
data <- lapply(1:n_datasets, function(i) {
  dataset <- datasets[i]
  parameters <- list(
    Dataset = dataset,
    PopGenUserName = "kevin.soandso@some.sci.org",
    PopulationSize = 100
  )
  if (dataset == "NDNS") {
    # See https://discover.ukdataservice.ac.uk/series/?sn=2000033#abstract
    parameters$AgeLower <- 2
    parameters$AgeUpper <- 4
    parameters$BodyMassIndexLower <- 10
    parameters$HeightLower <- 80
    parameters$HeightUpper <- 120
  }
  parameters <- apply_defaults(parameters)
  cat(paste("Sampling ", dataset, "\\n", sep = ""))
  # call service with no activity feedback
  pop_gen_out <- generate_population(parameters, silent = T)
  if (generate_failed(pop_gen_out)) {
    # show error messages now
    print(summary(pop_gen_out))
    stop()
  }
  if (i < n_datasets) {
    # service is throttled - must pause here
    cat("Waiting for service to resume...\\n")
    Sys.sleep(pop_gen_out$wait_for)
  }
  individuals <- extract_individuals(pop_gen_out)
  return(individuals)
})
names(data) <- datasets
}
\references{
McNally K., Cotton R., Hogg A., Loizou G. (2014). PopGen: a virtual human population generator. \emph{Toxicology} 315 70-85. 10.1016/j.tox.2013.07.009
\href{https://doi.org/10.1016/j.tox.2013.07.009}{[Cross ref]}
}
