% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{store_proxy_settings}
\alias{store_proxy_settings}
\title{Proxy Server Configuration}
\usage{
store_proxy_settings(url, port = NA_integer_, username = NA_character_,
  password = NA_character_, auth = "basic")
}
\arguments{
\item{url}{Location of proxy}

\item{port}{Location of proxy}

\item{username}{Login details for proxy, if needed}

\item{password}{Login details for proxy, if needed}

\item{auth}{Type of HTTP authentication to use. Should be one of the following: basic, digest, digest_ie, gssnegotiate, ntlm, any.}
}
\description{
Set up configuration when internet access is via a proxy server.
}
\details{
The configuration is stored as encrypted data in the user's home directory.

Refer to \code{\link[httr]{use_proxy}} for more information.
}
\examples{
# set up access to proxy server requiring authentication
store_proxy_settings("my_corporate_proxy_server", 8080, "kevin", "pass123!")
}
